Fix camera bug and device list and diet managment done.

+++ app/src/main/java/com/kau/smartbutler/model/Diet.kt
@@ -3,9 +3,9 @@ package com.kau.smartbutler.model
 import io.realm.RealmObject
 
 open class Diet(
-        var year: Int = 0,
-        var month: Int = 0,
-        var date: Int = 0,
+        var year: Long = 0,
+        var month: Long = 0,
+        var date: Long = 0,
         var carbo: Float = 0.toFloat(),
         var protein: Float = 0.toFloat(),
         var fat: Float = 0.toFloat()
